requirements.txt
Django
django-ckeditor-5
django-bootstrap-v5
django-mptt
django-ckeditor